# [generator:regions] Speedup locality index building: parallel regions stripping