--- a/base/thread_pool_computational.hpp
+++ b/base/thread_pool_computational.hpp
@@ -9,6 +9,7 @@
 #include <memory>
 #include <queue>
 #include <thread>
+#include <vector>
 
 namespace base
 {
@@ -100,6 +101,23 @@ public:
     m_condition.notify_one();
   }
 
+  // Submit min(|workersCountHint|, Size()) tasks and wait completions.
+  // func - task to be performed.
+  // Warning: If the thread pool is stopped then the call will be ignored.
+  template <typename F, typename... Args>
+  void PerformParallelWorks(F && func, size_t workersCountHint)
+  {
+    size_t const workersCount = std::min(std::max(size_t{1}, workersCountHint), Size());
+
+    std::vector<std::future<void>> workers{};
+    workers.reserve(workersCount);
+    for (size_t i = 0; i < workersCount; ++i)
+      workers.push_back(Submit(func));
+
+    for (auto & worker : workers)
+      worker.wait();
+  }
+
   // Stop a ThreadPool.
   // Removes the tasks that are not yet started from the queue.
   // Unlike the destructor, this function does not wait for all runnables to complete:
@@ -124,6 +142,8 @@ public:
     m_condition.notify_all();
     m_joiner.Join();
   }
+
+  size_t Size() const noexcept { return m_threads.size(); }
 
 private:
   void Worker()