Ускорение генерации regions_index.locidx.
Некоторое регионы требую большого объёма вычислений. Такие объекты обрабатываются параллельно.

Сам по себе PR ускоряет генерацию regions_index.locidx лишь 3m:30s (27m vs 30m:30s), но совместно с PR #78 генерация сокращается в 4 раза (7m:15s vs 30m:30s)